# database/config/docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:7001",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/searchdata/save": {
      "post": {
        "tags": [
          "search-data-controller"
        ],
        "summary": "Save a search data object",
        "description": "Saves a new search data and returns the saved search data.",
        "operationId": "save",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchDataDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SearchDataDto"
                }
              }
            }
          }
        }
      }
    },
    "/searchdata/save-all": {
      "post": {
        "tags": [
          "search-data-controller"
        ],
        "summary": "Save multiple search data objects",
        "description": "Saves a list of search data objects and returns the list of saved objects.",
        "operationId": "save_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/SearchDataDto"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchDataDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/searchdata/get/{id}": {
      "get": {
        "tags": [
          "search-data-controller"
        ],
        "summary": "Retrieve a search data object by its ID",
        "description": "Retrieves a search data object based on its ID. Yields HttpStatus.NOT_FOUND if no search data object was found.",
        "operationId": "getSearchDataById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchDataDto"
                }
              }
            }
          }
        }
      }
    },
    "/searchdata/get-for-drawing/{id}": {
      "get": {
        "tags": [
          "search-data-controller"
        ],
        "summary": "Retrieve search data by its associated drawing ID",
        "description": "Retrieves a list of search data objects associated with a specific drawing ID.Yields HttpStatus.NOT_FOUND if no search data object was found.",
        "operationId": "getSearchDataByDrawingId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchDataDto"
                }
              }
            }
          }
        }
      }
    },
    "/searchdata/get-all": {
      "get": {
        "tags": [
          "search-data-controller"
        ],
        "summary": "Retrieve all search data",
        "description": "Retrieves a list of all search data objects. Yields an empty list if no search data objects were found.",
        "operationId": "getAllSearchData",
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchDataDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/searchdata/delete/{id}": {
      "delete": {
        "tags": [
          "search-data-controller"
        ],
        "summary": "Delete a search data object by its ID",
        "description": "Deletes a search data object based on its ID.",
        "operationId": "deleteSearchDataById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/searchdata/delete-for-drawing/{id}": {
      "delete": {
        "tags": [
          "search-data-controller"
        ],
        "summary": "Delete search data by its associated drawing ID",
        "description": "Deletes search data associated with a specific drawing ID.",
        "operationId": "deleteSearchDataByDrawingId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "SearchDataDto": {
        "type": "object",
        "properties": {
          "searchdata_id": {
            "type": "integer",
            "format": "int32"
          },
          "drawing_id": {
            "type": "integer",
            "format": "int32"
          },
          "shape": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "float"
            }
          },
          "material": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "general_tolerances": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "surfaces": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "gdts": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "threads": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "outer_dimensions": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "float"
            }
          },
          "search_vector": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "float"
            }
          },
          "part_number": {
            "type": "string"
          },
          "ocr_text": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "runtime_text": {
            "type": "string"
          }
        }
      },
...
    }
  }
}